Add initial TabNet implementation with key components
Implemented the foundational components of the TabNet model, including the encoder, decoder, and key modules such as GLU, GBN, and attentive transformers. This commit also includes embedding generation and the overall TabNet architecture to enable feature processing and training workflows.
