+++ pytorch_tabnet/tab_network/embedding_generator.py
@@ -1,0 +1,101 @@
+"""EmbeddingGenerator class module."""
+
+from typing import List, Union
+
+import torch
+
+
+class EmbeddingGenerator(torch.nn.Module):
+    """Classical embeddings generator."""
+
+    def __init__(
+        self,
+        input_dim: int,
+        cat_dims: List[int],
+        cat_idxs: List[int],
+        cat_emb_dims: Union[List[int], int],
+        group_matrix: torch.Tensor,
+    ):
+        """This is an embedding module for an entire set of features.
+
+        Parameters
+        ----------
+        input_dim : int
+            Number of features coming as input (number of columns)
+        cat_dims : list of int
+            Number of modalities for each categorial features
+            If the list is empty, no embeddings will be done
+        cat_idxs : list of int
+            Positional index for each categorical features in inputs
+        cat_emb_dim : list of int
+            Embedding dimension for each categorical features
+            If int, the same embedding dimension will be used for all categorical features
+        group_matrix : torch matrix
+            Original group matrix before embeddings
+
+        """
+        super(EmbeddingGenerator, self).__init__()
+
+        if cat_dims == [] and cat_idxs == []:
+            self.skip_embedding = True
+            self.post_embed_dim = input_dim
+            self.embedding_group_matrix = group_matrix.to(group_matrix.device)
+            return
+        else:
+            self.skip_embedding = False
+
+        if isinstance(cat_emb_dims, int):
+            cat_emb_dims = [cat_emb_dims] * len(cat_dims)
+
+        self.post_embed_dim = int(input_dim + sum(cat_emb_dims) - len(cat_emb_dims))
+
+        self.embeddings = torch.nn.ModuleList()
+
+        for cat_dim, emb_dim in zip(cat_dims, cat_emb_dims, strict=False):
+            self.embeddings.append(torch.nn.Embedding(cat_dim, emb_dim))
+
+        # record continuous indices
+        self.continuous_idx = torch.ones(input_dim, dtype=torch.bool)
+        self.continuous_idx[cat_idxs] = 0
+
+        # update group matrix
+        n_groups = group_matrix.shape[0]
+        self.embedding_group_matrix = torch.empty((n_groups, self.post_embed_dim), device=group_matrix.device)
+        for group_idx in range(n_groups):
+            post_emb_idx = 0
+            cat_feat_counter = 0
+            for init_feat_idx in range(input_dim):
+                if self.continuous_idx[init_feat_idx] == 1:
+                    # this means that no embedding is applied to this column
+                    self.embedding_group_matrix[group_idx, post_emb_idx] = group_matrix[group_idx, init_feat_idx]  # noqa
+                    post_emb_idx += 1
+                else:
+                    # this is a categorical feature which creates multiple embeddings
+                    n_embeddings = cat_emb_dims[cat_feat_counter]
+                    self.embedding_group_matrix[group_idx, post_emb_idx : post_emb_idx + n_embeddings] = (
+                        group_matrix[group_idx, init_feat_idx] / n_embeddings
+                    )  # noqa
+                    post_emb_idx += n_embeddings
+                    cat_feat_counter += 1
+
+    def forward(self, x: torch.Tensor) -> torch.Tensor:
+        """Apply embeddings to inputs
+        Inputs should be (batch_size, input_dim)
+        Outputs will be of size (batch_size, self.post_embed_dim).
+        """
+        if self.skip_embedding:
+            # no embeddings required
+            return x
+
+        cols = []
+        cat_feat_counter = 0
+        for feat_init_idx, is_continuous in enumerate(self.continuous_idx):
+            # Enumerate through continuous idx boolean mask to apply embeddings
+            if is_continuous:
+                cols.append(x[:, feat_init_idx].float().view(-1, 1))
+            else:
+                cols.append(self.embeddings[cat_feat_counter](x[:, feat_init_idx].long()))
+                cat_feat_counter += 1
+        # concat
+        post_embeddings = torch.cat(cols, dim=1)
+        return post_embeddings
