+++ pytorch_tabnet/tab_network/tabnet_encoder.py
@@ -1,0 +1,195 @@
+"""TabNetEncoder class module."""
+
+from typing import List, Optional, Union
+
+import torch
+from torch.nn import BatchNorm1d, Linear, ReLU
+
+from .attentive_transformer import AttentiveTransformer
+
+# from pytorch_tabnet.tab_network.feat_transformer import FeatTransformer
+# from pytorch_tabnet.tab_network.attentive_transformer import AttentiveTransformer
+from .feat_transformer import FeatTransformer
+
+
+class TabNetEncoder(torch.nn.Module):
+    def __init__(
+        self,
+        input_dim: int,
+        output_dim: Union[int, List[int]],
+        n_d: int = 8,
+        n_a: int = 8,
+        n_steps: int = 3,
+        gamma: float = 1.3,
+        n_independent: int = 2,
+        n_shared: int = 2,
+        epsilon: float = 1e-15,
+        virtual_batch_size: int = 128,
+        momentum: float = 0.02,
+        mask_type: str = "sparsemax",
+        group_attention_matrix: Optional[torch.Tensor] = None,
+    ):
+        """Defines main part of the TabNet network without the embedding layers.
+
+        Parameters
+        ----------
+        input_dim : int
+            Number of features
+        output_dim : int or list of int for multi task classification
+            Dimension of network output
+            examples : one for regression, 2 for binary classification etc...
+        n_d : int
+            Dimension of the prediction  layer (usually between 4 and 64)
+        n_a : int
+            Dimension of the attention  layer (usually between 4 and 64)
+        n_steps : int
+            Number of successive steps in the network (usually between 3 and 10)
+        gamma : float
+            Float above 1, scaling factor for attention updates (usually between 1.0 to 2.0)
+        n_independent : int
+            Number of independent GLU layer in each GLU block (default 2)
+        n_shared : int
+            Number of independent GLU layer in each GLU block (default 2)
+        epsilon : float
+            Avoid log(0), this should be kept very low
+        virtual_batch_size : int
+            Batch size for Ghost Batch Normalization
+        momentum : float
+            Float value between 0 and 1 which will be used for momentum in all batch norm
+        mask_type : str
+            Either "sparsemax" or "entmax" : this is the masking function to use
+        group_attention_matrix : torch matrix
+            Matrix of size (n_groups, input_dim), m_ij = importance within group i of feature j
+
+        """
+        super(TabNetEncoder, self).__init__()
+        self.input_dim = input_dim
+        self.output_dim = output_dim
+        self.is_multi_task = isinstance(output_dim, list)
+        self.n_d = n_d
+        self.n_a = n_a
+        self.n_steps = n_steps
+        self.gamma = gamma
+        self.epsilon = epsilon
+        self.n_independent = n_independent
+        self.n_shared = n_shared
+        self.virtual_batch_size = virtual_batch_size
+        self.mask_type = mask_type
+        self.initial_bn = BatchNorm1d(self.input_dim, momentum=0.01)
+        self.group_attention_matrix = group_attention_matrix
+
+        if self.group_attention_matrix is None:
+            # no groups
+            self.group_attention_matrix = torch.eye(self.input_dim)
+            self.attention_dim = self.input_dim
+        else:
+            self.attention_dim = self.group_attention_matrix.shape[0]
+
+        if self.n_shared > 0:
+            shared_feat_transform = torch.nn.ModuleList()
+            for i in range(self.n_shared):
+                if i == 0:
+                    shared_feat_transform.append(Linear(self.input_dim, 2 * (n_d + n_a), bias=False))
+                else:
+                    shared_feat_transform.append(Linear(n_d + n_a, 2 * (n_d + n_a), bias=False))
+
+        else:
+            shared_feat_transform = None
+
+        self.initial_splitter = FeatTransformer(
+            self.input_dim,
+            n_d + n_a,
+            shared_feat_transform,
+            n_glu_independent=self.n_independent,
+            virtual_batch_size=self.virtual_batch_size,
+            momentum=momentum,
+        )
+
+        self.feat_transformers = torch.nn.ModuleList()
+        self.att_transformers = torch.nn.ModuleList()
+
+        for _step in range(n_steps):
+            transformer = FeatTransformer(
+                self.input_dim,
+                n_d + n_a,
+                shared_feat_transform,
+                n_glu_independent=self.n_independent,
+                virtual_batch_size=self.virtual_batch_size,
+                momentum=momentum,
+            )
+            attention = AttentiveTransformer(
+                n_a,
+                self.attention_dim,
+                group_matrix=group_attention_matrix,
+                virtual_batch_size=self.virtual_batch_size,
+                momentum=momentum,
+                mask_type=self.mask_type,
+            )
+            self.feat_transformers.append(transformer)
+            self.att_transformers.append(attention)
+
+    def forward(self, x: torch.Tensor, prior: Optional[torch.Tensor] = None) -> tuple[List[torch.Tensor], torch.Tensor]:
+        x = self.initial_bn(x)
+
+        bs = x.shape[0]  # batch size
+        if prior is None:
+            prior = torch.ones((bs, self.attention_dim), device=x.device)
+
+        M_loss = torch.tensor(0.0, device=x.device)
+        att = self.initial_splitter(x)[:, self.n_d :]
+        steps_output = []
+        for step in range(self.n_steps):
+            M = self.att_transformers[step](prior, att)
+            M_loss += torch.mean(torch.sum(torch.mul(M, torch.log(M + self.epsilon)), dim=1))
+            # update prior
+            prior = torch.mul(self.gamma - M, prior)
+            # output
+            M_feature_level = torch.matmul(M, self.group_attention_matrix)
+            masked_x = torch.mul(M_feature_level, x)
+            out = self.feat_transformers[step](masked_x)
+            d = ReLU()(out[:, : self.n_d])
+            steps_output.append(d)
+            # update attention
+            att = out[:, self.n_d :]
+
+        M_loss /= self.n_steps
+        return steps_output, M_loss
+
+    def forward_masks(self, x: torch.Tensor) -> tuple[torch.Tensor, dict]:
+        """Compute and return feature masks for each step.
+
+        Parameters
+        ----------
+        x : torch.Tensor
+            Input tensor.
+
+        Returns
+        -------
+        tuple[torch.Tensor, dict]
+            Tuple of (explanation mask, step-wise masks dictionary).
+
+        """
+        x = self.initial_bn(x)
+        bs = x.shape[0]  # batch size
+        prior = torch.ones((bs, self.attention_dim)).to(x.device)
+        M_explain = torch.zeros(x.shape).to(x.device)
+        att = self.initial_splitter(x)[:, self.n_d :]
+        masks = {}
+
+        for step in range(self.n_steps):
+            M = self.att_transformers[step](prior, att)
+            M_feature_level = torch.matmul(M, self.group_attention_matrix)
+            masks[step] = M_feature_level
+            # update prior
+            prior = torch.mul(self.gamma - M, prior)
+            # output
+            masked_x = torch.mul(M_feature_level, x)
+            out = self.feat_transformers[step](masked_x)
+            d = ReLU()(out[:, : self.n_d])
+            # explain
+            step_importance = torch.sum(d, dim=1)
+            M_explain += torch.mul(M_feature_level, step_importance.unsqueeze(dim=1))
+            # update attention
+            att = out[:, self.n_d :]
+
+        return M_explain, masks
